Do not reinstall

+++ sbotools/setup.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-wget https://pink-mist.github.io/sbotools/downloads/sbotools-2.7-noarch-1_SBo.tgz -O /tmp/sbotools-2.7-noarch-1_SBo.tgz &&
-  installpkg /tmp/sbotools-2.7-noarch-1_SBo.tgz
+if [ -f "/var/log/packages/sbotools-2.7-noarch-1_SBo" ] ; then
+  wget https://pink-mist.github.io/sbotools/downloads/sbotools-2.7-noarch-1_SBo.tgz -O /tmp/sbotools-2.7-noarch-1_SBo.tgz &&
+    installpkg /tmp/sbotools-2.7-noarch-1_SBo.tgz
+fi
 
 mkdir -p /etc/sbotools
   cp sbotools/sbotools.conf /etc/sbotools
