+++ slackpkg+/setup.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-wget "https://ufpr.dl.sourceforge.net/project/slackpkgplus/slackpkg+-1.7.0-noarch-10mt.txz" -O "/tmp/slackpkgplus+-1.7.0-noarch.10mt.txz" &&
-  upgradepkg --reinstall --install-new "/tmp/slackpkgplus+-1.7.0-noarch.10mt.txz"
+if [ -f "/var/log/packages/slackpkg+-1.7.0-noarch-10mt" ] ; then
+  wget "https://ufpr.dl.sourceforge.net/project/slackpkgplus/slackpkg+-1.7.0-noarch-10mt.txz" -O "/tmp/slackpkgplus+-1.7.0-noarch.10mt.txz" &&
+    upgradepkg --reinstall --install-new "/tmp/slackpkgplus+-1.7.0-noarch.10mt.txz"
+fi
 
 for file in blacklist greylist slackpkg.conf slackpkgplus.conf ; do
   cp "slackpkg+/$file" /etc/slackpkg
